Removed Jam's shitfuckery
This mod is called Liberty, Egality, Fraternity: legacy of '48. It is based of a world where mnay of the 1848 revolutions succeed, which leads to a weaker europe and a vey different world. This mod was produced with the help and assisstance of RanWithAPlan, (put your name here) and we hope you'll enjoy it.
Have fun,
    The Dev Team